+++ exporters/otlp/internal/transform/attribute.go
@@ -15,7 +15,7 @@
 package transform
 
 import (
-	commonpb "go.opentelemetry.io/otel/internal/opentelemetry-proto-gen/common/v1"
+	commonpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/common/v1"
 
 	"go.opentelemetry.io/otel/api/kv"
 	"go.opentelemetry.io/otel/sdk/resource"