+++ exporters/otlp/otlp_integration_test.go
@@ -25,8 +25,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	commonpb "go.opentelemetry.io/otel/internal/opentelemetry-proto-gen/common/v1"
-	metricpb "go.opentelemetry.io/otel/internal/opentelemetry-proto-gen/metrics/v1"
+	commonpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/common/v1"
+	metricpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/metrics/v1"
 
 	"go.opentelemetry.io/otel/api/kv"
 	"go.opentelemetry.io/otel/api/metric"